Fetch documnet data from selected item
